6-size.c

+++ 0x00-hello_world/6-size.c
@@ -10,11 +10,11 @@ char c;
 int i;
 long lg;
 long long int x;
-unsigned float f;
+float f;
 printf("Size of a char: %d byte(s)\n", sizeof(c));
 printf("Size of an int: %d byte(s)\n", sizeof(i));
 printf("Size of a long int: %d byte(s)\n", sizeof(lg));
 printf("Size of a long long int: %d byte(s)\n", sizeof(x));
-printf("Size of a float: %1.0f byte(s)\n", sizeof(f));
+printf("Size of a float: %f byte(s)\n", sizeof(f));
 return (0);
 }
